asl-recognition-web-main/requirements.txt
flask
flask-cors
opencv-python
SpeechRecognition
openai
python-dotenv
Pillow
pygame
sounddevice
numpy
pyaudio
mediapipe
elevenlabs>=2.0.0